requirements/test.txt
#
# This file is autogenerated by pip-compile
# Make changes in test.in, then run this to update:
#
#    pip-compile test.in
#
click==5.1                # via pip-tools
coverage==4.0
crispy-forms-foundation==0.5.2
defusedxml==0.4.1         # via python3-openid
django-allauth==0.23.0
django-appconf==1.0.1     # via django-compressor
django-autoslug==1.8.0
django-bower==5.0.4
django-braces==1.8.1
django-cacheops==2.4.1
django-compressor==1.5
django-crispy-forms==1.5.2
django-environ==0.3.0
django-filter==0.11.0
django-floppyforms==1.5.2
django-grappelli==2.7.1
django-model-utils==2.3.1
django-redis==4.2.0
django-secure==1.0.1
django-test-plus==1.0.9
django==1.8.4
djangorestframework==3.2.3
docutils==0.12
factory-boy==2.5.2
first==2.0.1              # via pip-tools
flake8==2.4.1
funcy==1.6                # via django-cacheops
mccabe==0.3.1             # via flake8
msgpack-python==0.4.6     # via django-redis
oauthlib==1.0.3           # via requests-oauthlib
pep8==1.5.7               # via flake8
pillow==2.9.0
pip-tools==1.1.4
psycopg2==2.6.1
pyflakes==0.8.1           # via flake8
python3-openid==3.0.7     # via django-allauth
pytz==2015.6
redis==2.10.3
requests-oauthlib==0.5.0  # via django-allauth
requests==2.7.0           # via django-allauth, requests-oauthlib
six==1.9.0                # via django-appconf, django-bower, django-braces, django-cacheops, pip-tools, unicode-slugify
unicode-slugify==0.1.3
unidecode==0.4.18         # via unicode-slugify
whitenoise==2.0.4
